refactor: changed ticket type hook api name
